remove SettingsCoordinator
